dispFunc: debugging

+++ source/dispFunctions.h
@@ -10,6 +10,6 @@
 
 using namespace arma;
 
-void finiteAppliedDisp(vec & time, vec & disp, vec & eps, int nsteps, int time_tot, double strainrate, double h);	
+void finiteAppliedDisp(vec & time, vec & disp, vec & eps, int nsteps, double time_tot, double strainrate, double h);	
 
-void shpbAppliedDisp(vec & time, vec & disp, vec & eps, int nsteps, int time_tot, double strainrate, double h);
+void shpbAppliedDisp(vec & time, vec & disp, vec & eps, int nsteps, double time_tot, double strainrate, double h);
